changed katex location to Message.svelte
