+++ .env
@@ -1,3 +1,5 @@
 # Uncomment the following line to use the mocked API
 #VITE_API_BASE=http://localhost:5174
+
+# Whether to render latex
 VITE_RENDER_LATEX=true
