+++ src/lib/Messages.svelte
@@ -3,12 +3,30 @@
     import SvelteMarkdown from 'svelte-markdown'
     import type { Message, Model, Usage } from './Types.svelte'
 
+    import {marked} from "marked";
+    import markedKatex from "marked-katex-extension";
+    import '../katex.min.css';
+
+    // whether to render latex; default is true
+    const renderLatexFlag = JSON.parse(import.meta.env.VITE_RENDER_LATEX) || false
+
     // Marked options
     const markedownOptions = {
       gfm: true, // Use GitHub Flavored Markdown
       breaks: true, // Enable line breaks in markdown
       mangle: false // Do not mangle email addresses
     }
+
+    // Katex rendering
+    const renderLatex = (text: string): string => {
+        if (renderLatexFlag == true) {
+        const katexOptions = { throwOnError: false };
+        marked.use(markedKatex(katexOptions));
+        return marked(text);
+        }
+
+        return text;
+    };
 
     export let messages : Message[]
     export let input: HTMLTextAreaElement
@@ -61,7 +79,7 @@
   {:else}
     <article class="message is-success assistant-message">
       <div class="message-body content">
-        <SvelteMarkdown source={message.content} options={markedownOptions} renderers={{ code: Code, html: Code }}/>
+        <SvelteMarkdown source={renderLatex(message.content)} options={markedownOptions} renderers={{ code: Code, html: Code }}/>
         {#if message.usage}
           <p class="is-size-7">
             This message was generated on <em>{message.model || defaultModel}</em> using <span class="has-text-weight-bold">{message.usage.total_tokens}</span>
